Restructure folder as intended.
