+++ src/Microsoft.Performance.SDK.Tests/TestClasses/StubTableProvider.cs
@@ -5,15 +5,15 @@ using System;
 using System.Collections.Generic;
 using Microsoft.Performance.SDK.Processing;
 
-namespace Microsoft.Performance.SDK.Tests
+namespace Microsoft.Performance.SDK.Tests.TestClasses
 {
-    public class FakeTableProvider
+    public class StubTableProvider
         : IProcessingSourceTableProvider
     {
-        public FakeTableProvider()
+        public StubTableProvider()
         {
-            this.DiscoverCalls = new List<ITableConfigurationsSerializer>();
-            this.DiscoverReturnValue = Array.Empty<TableDescriptor>();
+            DiscoverCalls = new List<ITableConfigurationsSerializer>();
+            DiscoverReturnValue = Array.Empty<TableDescriptor>();
         }
 
         public List<ITableConfigurationsSerializer> DiscoverCalls { get; }
@@ -22,8 +22,8 @@ namespace Microsoft.Performance.SDK.Tests
 
         public IEnumerable<TableDescriptor> Discover(ITableConfigurationsSerializer tableConfigSerializer)
         {
-            this.DiscoverCalls.Add(tableConfigSerializer);
-            return this.DiscoverReturnValue;
+            DiscoverCalls.Add(tableConfigSerializer);
+            return DiscoverReturnValue;
         }
     }
 }
