+++ src/Microsoft.Performance.SDK.Tests/TestClasses/StubTables.cs
